Adjusted generic pagination return value.
