--- Eskon.Core/Response/Paginated.cs
@@ -1,20 +1,0 @@
-﻿namespace Eskon.Core.Response
-{
-    public class Paginated<T>
-    {
-        public List<T> Data { get; }
-        public int PageNumber { get; }
-        public int PageSize { get; }
-        public int TotalPages { get; }
-        public int TotalRecords { get; }
-
-        public Paginated(List<T> data, int pageNumber, int pageSize, int totalRecords)
-        {
-            Data = data;
-            PageNumber = pageNumber;
-            PageSize = pageSize;
-            TotalRecords = totalRecords;
-            TotalPages = (int)Math.Ceiling(totalRecords / (double)pageSize);
-        }
-    }
-}
