+++ Eskon.Infrastructure/Generics/IGenericRepositoryAsync.cs
@@ -1,4 +1,5 @@
-﻿using Eskon.Domian.Entities;
+﻿using Eskon.Domain.Utilities;
+using Eskon.Domian.Entities;
 using Microsoft.EntityFrameworkCore.Storage;
 using System.Linq.Expressions;
 
@@ -12,9 +13,9 @@ namespace Eskon.Infrastructure.Generics
         Task UpdateRangeAsync(ICollection<T> entities);
         Task<List<T>> GetAllAsync();
         Task<List<T>> GetFilteredAsync(Expression<Func<T, bool>>? filter = null);
-        Task<List<T>> GetPageAsync(int pageNumber = 1, int ItemsPerPage = 10);
-        Task<List<T>> GetPaginatedAsync(int pageNumber = 1, int ItemsPerPage = 10, Expression<Func<T, bool>>? filter = null);
-        Task<List<T>> GetPaginatedSortedAsync<TKey>(Expression<Func<T, TKey>> sort, bool asc, int pageNumber = 1, int itemsPerPage = 10, Expression<Func<T, bool>>? filter = null);
+        Task<Paginated<T>> GetPageAsync(int pageNumber = 1, int ItemsPerPage = 10);
+        Task<Paginated<T>> GetPaginatedAsync(int pageNumber = 1, int ItemsPerPage = 10, Expression<Func<T, bool>>? filter = null);
+        Task<Paginated<T>> GetPaginatedSortedAsync<TKey>(Expression<Func<T, TKey>> sort, bool asc, int pageNumber = 1, int itemsPerPage = 10, Expression<Func<T, bool>>? filter = null);
         Task<int> GetTotalCount();
         Task<T> GetByIdAsync(object id);
         Task DeleteAsync(T entity);
